Redesign virtual-keyboard integration: native commands, events, declarative --keyboard-height
The renamed tauri-plugin-virtual-keyboard (was -padding) no longer resizes
the webview. It maintains a per-frame --keyboard-height CSS variable, emits
willShow/willHide/didShow/didHide/change events feeding signalium-signal
keyboard state, and exposes native hide/show commands. The app pads
MobileLayout by the variable (reproducing resize semantics declaratively),
bridges signals with useSignal, and drops the viewport-inference tracker
(keyboard.svelte.ts) and the pulse/pinSlot machinery in
render-below-keyboard.
