+++ pnpm-workspace.yaml
@@ -14,6 +14,8 @@ allowBuilds:
   geckodriver: true
   # Builds dist-js via its `prepare` script when installed from the git fork.
   "@gbyte/tauri-plugin-ios-photos": true
+  # Builds dist via its `prepare` script when installed from the git repo.
+  tauri-plugin-virtual-keyboard: true
 
 patchedDependencies:
   konsta@5.0.6: patches/konsta@5.0.6.patch
