+++ src-tauri/capabilities/main.json
@@ -16,6 +16,7 @@
 		"sharekit:allow-share-file",
 		"fs:allow-appcache-write-recursive",
 		"mailto:default",
-		"core:path:default"
+		"core:path:default",
+		"virtual-keyboard:default"
 	]
 }
