+++ ui/package.json
@@ -43,6 +43,7 @@
 		"tailwindcss": "^4.1.18",
 		"tauri-plugin-android-fs-api": "^28.2.1",
 		"tauri-plugin-view-api": "^0.0.5",
+		"tauri-plugin-virtual-keyboard": "github:dash-chat/tauri-plugin-virtual-keyboard#2fc30fe69d334a0dcec62257b2c5d2b335c04f72",
 		"vite-plugin-babel": "^1.3.2"
 	},
 	"devDependencies": {
