+++ ui/src/routes/+layout.svelte
@@ -42,7 +42,7 @@
 	import { applyDarkMode } from '$lib/utils/theme';
 	import { showToast } from '$lib/utils/toasts';
 	import { isIos, isMobile, isTauriEnv } from '$lib/utils/environment';
-	import { trackKeyboardHeight } from '$lib/utils/keyboard.svelte';
+	import { trackKeyboardHeight } from 'tauri-plugin-virtual-keyboard';
 	import { forwardConsoleToTauriLog } from '$lib/utils/logs';
 	import {
 		listenForDeepLinks,
